fixed dumb mistake

+++ System Libs/iolib/Program.cs
@@ -41,7 +41,7 @@ namespace io
         
         public static int FileSize(string path)
         {
-            return new System.IO.FileInfo(path).Length;
+            return (int)(new System.IO.FileInfo(path).Length);
         }
 
         public static string[] GetFilesInDirectory(string path)
